CRUD for assets

+++ src/main/java/com/aleix/XposeAPI/service/AssetService.java
@@ -1,9 +1,12 @@
 package com.aleix.XposeAPI.service;
 
+import com.aleix.XposeAPI.controller.FileUploadController;
 import com.aleix.XposeAPI.model.Asset;
 import com.aleix.XposeAPI.repository.AssetRepository;
 import com.aleix.XposeAPI.specification.ArtistSpecifications;
+import com.aleix.XposeAPI.specification.AssetSpecifications;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,9 +15,11 @@ import java.util.Optional;
 public class AssetService {
 
     private final AssetRepository assetRepository;
+    private final FileUploadService fileUploadService;
 
-    public AssetService(AssetRepository assetRepository) {
+    public AssetService(AssetRepository assetRepository, FileUploadService fileUploadService) {
         this.assetRepository = assetRepository;
+        this.fileUploadService = fileUploadService;
     }
 
     public List<Asset> getAllAssets() {
@@ -25,7 +30,8 @@ public class AssetService {
         return assetRepository.findById(id);
     }
 
-    public Asset createAsset(Asset asset) {
+    public Asset createAsset(Asset asset, MultipartFile file) throws Exception {
+        asset.setUrl(fileUploadService.uploadFile(file));
         return assetRepository.save(asset);
     }
 
@@ -42,7 +48,18 @@ public class AssetService {
     }
 
     public boolean deleteAsset(Long id) {
-        if (assetRepository.existsById(id)) {
+        if (assetRepository.existsById(id) && assetRepository.findById(id).isPresent()) {
+
+            Asset asset = assetRepository.findById(id).get();
+
+            try {
+
+                fileUploadService.deleteObject(asset.getUrl());
+
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+
             assetRepository.deleteById(id);
             return true;
         }
@@ -65,8 +82,7 @@ public class AssetService {
                 .toList();
     }
 
-    //TODO ADD SPECIFICATION
-    /*public List<Asset> filterArtists (String name, String surname, String artisticName){
-        return assetRepository.findAll(ArtistSpecifications.filterArtists(name, surname, artisticName));
-    }*/
+    public List<Asset> filterAssets (String name, String type, Boolean active, String artistId, String serieId){
+        return assetRepository.findAll(AssetSpecifications.filterAssets(name, type, active, artistId,  serieId));
+    }
 }
